requirements.txt
Django
django-allauth
django-tailwind
django-htmx
django-crispy-forms
crispy-tailwind
python-dotenv
whitenoise
gunicorn
dj-database-url
ruff
openpyxl